requirements.txt
pandas
langchain
langchain-community
langchain-astradb
python-dotenv
langchain_google_genai
-e .